V 0.2

+++ WebApplication1/WebAppServer/Models/Alumno.cs
@@ -5,7 +5,7 @@
     {
         public string Nombre { get; set; }
         public int Id { get; set; }
-        public double Nota { get; set; }
+        public decimal Nota { get; set; }
         public int LU {  get; set; }
     }
 }
